requirements.txt
anthropic>=0.40.0
PyGithub>=2.8.0
requests==2.31.0
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
bandit==1.7.6
pylint==3.0.3
radon==6.0.1
gitpython==3.1.40
